Remove permission checking
Now we’re in our own application, we don’t need to check a user’s
specific role, as GDS-sso checks if a user can log in for us. We may
want to add more fine-grained permissions later, but for now, if a user
has access to Content Block Manager is enough.