Add functionality to import data
We have some JSON files (one for each model), which we intend to upload
to the root application directory. Once this is done, we can run this
task. The task runs in a transaction, so if there are any errors (such
as a file not being found, or a validation error), any changes are
rolled back