Reset primary key sequence
When importing the data, we include the IDs as well. This means that the
primary key sequence is now out of whack, and we get an error when
creating subsequent rows. This adds a call to the import to reset the
primary key sequence using [`reset_pk_sequence!`](https://apidock.com/rails/ActiveRecord/ConnectionAdapters/PostgreSQL/SchemaStatements/reset_pk_sequence%21),
which restores the sequence of the table’s primary key to the maximum
value.