Remove ContentBlockManager::ContentBlock namespace
The code we’ve ported over from Whitehall was located within
a `ContentBlockManager::ContentBlock` namespace. This was a
relic from its previous location in an “engine“ within Whitehall.
Now that the code is in its own application we remove this redundant
namespacing as it:

-  makes the code verbose and difficult to scan
- causes it to be nested in redundant sub-directories

This is a huge PR, but there are no changes in behaviour included.
It's purely a restructuring and simplification of the naming.

The benefit of making the code base much clearer outweighs the
disruption to the narrative flow of the git history. This improvement is
an important benefit of the "re-platforming" effort.

See:

- [ADR 3 in Whitehall][] (previous home, as an "engine")
- [ADR 3 in Content Block Manager][] (new dedicated home)

[ADR 3 in Whitehall]:
(https://github.com/alphagov/whitehall/blob/main/docs/content_block_manager/adr/0003-move-out-to-new-application.md)

[ADR 3 in Content Block Manager]:
(https://github.com/alphagov/content-block-manager/blob/main/docs/architecture/decisions/0003-move-out-to-new-application.md)