Initialize Turbo
This adds the relevant JS tags to the layout, as well as adds a listener
to reinitialize all the GOVUK modules when Turbo triggers a new page
load. I’ve also removed the `govuk_publishing_components` Govspeak
component, which was causing some issues - we don’t actually use it
anywhere, so is safe to remove.