Enable Turbo only where we want it
We've introduced the use of Rails' "Turbo drive" mechanism
for loading fragments of HTML for enhanced performance and
interactivity without introducing the overhead of custom JS
for Ajax requests. However, we don't want Turbo to highjack
all form POSTs as this plays havoc with our existing
behaviour, especially the handling of errors.

To disable Turbo by default it's recommended to set
`data-turbo="false"` [on the body tag][] but as we're using
the [admin layout component][] we don't have ready access to
the `body` tag.

Instead we set it on the `.govuk-main-wrapper` element.

And we enable Turbo in the specific places where we need it:

- the "Preview" button on the "group objects" and "select
  subschema" pages

- the "preview.turbo-stream" template which renders the
  "Close preview button"

[on the body tag]:
https://github.com/hotwired/turbo/issues/918#issuecomment-1545241471

[admin layout component]:
https://components.publishing.service.gov.uk/component-guide/layout_for_admin