Add methods to copy editions without specific details
Currently there’s no way to break blocks into their constituent parts
without including the main title of the block and excluding the title
of the part. I did consider if we might want to add this functionality
into the Tools gem, but this is only functionality that we need in
the application, so it’s less fiddly to do it here. This will allow us
to render specific parts of blocks to allow them to be reordered.