Enable Turbo in reordering
This adds `turbo: true` to the reordering buttons, so the reordering
feels a bit more snappy. We also wrap the component in a `turbo-frame`
to ensure the scroll position is retained.

Additionally, I’ve updated the Cucumber specs to ensure we wait until
the move has actually taken place before we click to save. To get this
to work we need to include the Playwright Matchers to make use of
Playwright’s auto-waiting functionality