Order subschemas by group_order for default_order
When fetching the default order, subschemas should be ordered by
the order defined in the config (via [`group_order`](https://github.com/alphagov/content-block-manager/blob/main/docs/configuration.md#schemasschema_namesubschemassubschema_namegroup_order)
if present).

We had the tests set up with this in mind, but they were passing by
accident, because the order we expected was linear. I’ve fixed the
tests, so the `group_order` defined in the test setups is no longer
linear, so we can confirm that subschemas are being ordered correctly.