Add failing test when blocks are created after reorder
If a block is created/edited, reordered, and then has a subsequent
item added to it, when the user goes to reorder the items, the new item
is missing, because it is not included in the order. This doesn’t matter
when rendering the block, as any missing items are appended to the end,
but this is not done when rendering the reorder component.