Update BDD tests to remove coupling from button text.
This change updates the existing Cucumber tests to remove the coupling
on the 'publish' or 'schedule' text on the form submit button. The
reason for this is that this is subject to change and may make the
tests more fragile.

I thought initially about changing this to just pick the button where
the type is 'submit' however on some pages we have multiple submit
buttons so this wouldn't work. (In hindsight this could possibly just
pick the first submit button it finds if it finds multiple.)

In the end I chose to select by the button class. This is also not
necessarily the best choice of selector however I feel like it's
unlikely to change any time soon. If needed I can add a comment to
this effect to explain the rationale.