Configure email_address embed_code as visible
This makes the contact.email_addresses.email_address field
render its embed code.

Against the actual "email address" we now have a link to:

{{embed:content_block_contact:my-contact/email_addresses/email/email_address}}

This is in addition to the existing embed code link for the
overall email address *block*:

{{embed:content_block_contact:my-contact/email_addresses/email}}

This is just to demonstrate the feature, as the utility of
this feature is still unclear.