Add button_text helper with logic for new blocks.
This adds a new helper which I called WorkflowHelper. I shied away
from using the EditionHelper as I feel like this doesn't relate to the
domain of Editions, it's more like view logic. As such this probably
could also be in a Presenter however I don't see many of those in this
codebase.

Added the logic for showing the 'Create' text based on existing method
in the Document.