Ensure buttons are rendered correctly when reordering
When rendering the buttons, we were still using the original `@order`
instance variable, which doesn’t include any newly added items, so the
`up` and `down` buttons were rendering incorrectly. We were also using
the initially set `order` when rendering the URLs for moving the items,
so I’ve fixed that too, as well as adding tests to ensure the buttons
are rendered correcltly with the right paths.