Update CSS and JS to point to the base path
When rendering the preview of a host document, we’re taking the HTML
source and replacing the content of the live block with the draft one.

I noticed that this doesn’t always render all of the styles, as we’re
relying on the CSS classes that exist in content block manager, so the
preview isn’t rendering 100% correctly.

This goes through the HTML header and appends the base path to each
piece of CSS and JS we encounter, ensuring all the CSS/JS in the preview
renders correctly.