Only show preview button for composed blocks
We initially added a preview button only in the grouped objects
workflow step, which only shows for composed blocks. Since then,
we also added the button at the review stage, which shows for all
block types.

As non-composed blocks aren’t embeddable as a single block, then
showing a preview of the block is irrelevant, so let’s hide the
link to the preview.