Style video relay service prefix textarea to single line
At present we use a Govspeak-enabled textarea as we understand
that users need to include an HTML link using markdown e.g.
`[Relay UK](https://www.relayuk.bt.com)`. But we
understand that it's not appropriate to offer a regular
textarea here as we need to communicate to users that the
content should fit on a single line.

We currently only support Govspeak on textareas so the fix
proposed is to style the Govspeak-enabled textarea to be a
single line. In this draft commit we:

- set the textarea height to govuk-spacing(7) which is 40px.
This is also the defined minimum-height of govuk textareas

- disable resizing of the textarea

However, in a desk review we questioned whether in addition
we'd need to disable the "return/enter" key to prevent the
user adding line breaks? Otherwise multiple line content can
be added but invisible. Furthermore, if multiline
content is then saved, only the first line will then be
visible when the form is rendered for editing.

An alternative might be to implement a Govspeak-enabled text
input but this is a more significant piece of work. Do we anticipate
other scenarios where we need a single line textarea?

Is it worth considering a two-line textarea? This might
have the advantage of:

- being equally effective at communicating that we expect a
  small amount of content, but

- also providing more functionality if the user does enter
  line breaks: they will be able to see the second line and
  have access to a minimal scroll bar.