Show embed code for each attribute
We're going to introduce a different mechanism for
controlling the visibility of embed codes. As:

1. the default position should be to display embed codes,
and

2. we recognise that certain fields are used internally as
flags to record whether or not a nested object is in use

we're going to replace `{Schema,EmbeddedSchema#embed_code_visible?}`
with `{Schema,EmbeddedSchema#hidden_field?}`.

In this commit, we do away with `#embed_code_visible` and
make all attrbiutes show their embed code.