Add #hidden_field? test to EmbeddedSchema and Schema
We're changing the logic around how embed codes are
displayed. Rather than controlling whether a given property
has its embed code displayed or not we're going to show ALL
embed codes, EXCEPT those for properties which are
configured as "hidden" fields.

These "hidden fields" are a class of attribute which does
not represent "content". For example, in a "Contact":

- `bsl_guidance.show`
- `call_charges.show_call_charges_info_url`
- `video_relay_service.show`
- `opening_hours.show_opening_hours`

These attributes are used as boolean (yes/no) flags, to record
whether in this instance of the block a certain element is to
be included or not.