Tests for error messages when user makes no selection
This changes an existing feature test and adds a similar new one to
test the error messages that are shown when a user doesn't make a
required selection.

I've changed the implementation of this to not use i18n as this can
lead to flaky tests. For example, in the previous commit I would have
broken the path to one of the translation strings; however, as the
test is using the same i18n locator it would have still passed as the
implementation and test would have both been broken equally.

I've also fixed the now-failing unit tests.