Improve styling of embed code 'flash' element
We now place the ephemeral "flash" element in a wrapper. By
giving this wrapper `position: relative` we can set the
flash to be absolutely positioned flush right with the edge
of the summary card and the "Copy code" link.

We set the embed code to display on a single line, to
improve readability.