Use JS to remove embed code from default block
As well as removing all the embed codes from embedded
objects we now remove the default block's embed code.

This is done using our Javascript `CopyEmbedCode` module to
remove the elements from the DOM after page load. In this way:

- if a user has JS disabled, they have access to the visible
embed codes which won't haven't removed by `CopyEmbedCode`

- if the user has JS enabled, they have access to the codes
via the JS-enabled "Copy code" links which i) copy the code
to the user's clipboard and ii) display the embed code for 2
seconds via an ephemeral "flash" element.

To achieve this neatly using the existing `CopyEmbedCode`
module, we relocate the embed code to the row above:

Before:
- row with default block (the JS module is bound to this row)
- row with embed code

After:
- row with default block, followed by embed code (JS module
  bound to this row)