Allow GA4 messages to be logged to the console
When we set `window.GOVUK.analyticsGa4.showDebug` to `true` we can see
the messages that would get sent to Google Analytics. This is annoying
to have to set each time a page loads, so let’s have an env var that
allows us to set it on each page load. This will also be useful in
feature tests