Make H2 tags unique in summary card, for screen readers.
The way I've achieved this is to call SummaryCardComponent with a
counter/index using the built in 'collection counter' functionality on
a rails ViewComponent. This relies on having a parameter in your
component with the same name as the 'collection parameter' appended
with `_count` and Rails handles the rest:

https://viewcomponent.org/guide/collections.html#collection-counter

This index is used to differentiate the `<h2>` tags by rendering the
count as invisible text inside the h2.

The index is grouped by the block_type so if there are 4 email
addresses followed by two telephone numbers, the index will go
in the order (1, 2, 3, 4, 1, 2).