Refactor steps code
This will allow us to get steps for a particular edition, so we can
confirm that GA4 messages have been set for each step. I’ve also
renamed the `Workflow::Steps` concern to `Workflow::HasSteps` so the
we can have a `Workflow::Steps` class to contain all the info about
steps.