Add a `populate_with_defaults` argument to form components
This checks if we should pre-populate a field with the default
values. If so, we return the default value if the field is blank,
if not, we return `nil`.

I’ve also made some tweaks to the test to make them a bit less
verbose, and only test what we are asserting in the test’s name.