Assert empty values don’t get repopulated with defaults
This updates the schema setup to make the `email_address`
subschema have a default value for `title` (as well as making
the `title` optional.

We then add a test to ensure the field doesn’t get repopulated
with a default value when it’s had the default value removed.

This will fail initially, but further commits will fix this.