Add awaiting_2i workflow state
This adds the new `awaiting_2i` workflow state only. There will be
other new states added and our existing state transitions will be
edited as part of this work but this will come in subsequent PRs.

As this will be feature flagged, the existing transitions need to be
maintained for now.