Migrate Workflow test
This is a biggie, but sets a lot of wheels in motion that should make
migrating furture integration tests easier. I’ve had to break the
tests up, as some of the embedded content tests required links to be
clicked, so weren’t request specs, so I’ve put them in a specific
feature test of their own.