Move require inside Rake task
At the moment, when we try and build the app for production using
Docker, we load all the Rake tasks, including the converter task.
As this requires `convert_to_rspec`, which in turn requires gems
that are not used in production, the application chokes because
it can’t find the gems. Moving the `require` statement to the task
ensures we only require the code when we run the Rake task, not
when we load it, which means we won’t try and load the code
in production.