Remove `environment` from Converter rake task
We suspect this is causing the Rake task to be eager loaded on
boot, and as the Rake task requires a script that requires dev/test
gems, this is causing the app to choke on boot.

I’ve also had to add an ignore rule for this file, as Rubocop’s
Rails rules require `:environment` to be included.