Ignore `parsers` directory when autoloading
This tells Zeitwerk to ignore this directory when attempting
to autoload code. As this is only used in dev, we don’t need
to to be required in prod, and this causes problems, as it
requires development gems.

We partially nicked the parser code from the `collections` app
(https://github.com/alphagov/collections/blob/53f903ab6499c63fb8889e4aab8ee4e7c8e384a7/lib/parsers/convert_to_rspec.rb)
but this wasn’t an issue for them, as the app was running Rails 6 at the
time, and [autoloading the lib dir wasn’t introduced until Rails 7.1][1]

[1]: https://blog.saeloun.com/2023/10/16/rails-7-1-introduces-autoload-lib/