Rename 'validate_review_page' update action to 'complete_workflow'
Each workflow step has a 'show action' and an 'update action'. The
update action for the review step was called validate_review_page
which I felt no longer accurately described the action taken at the
end of this step. This step now handles scheduling, publishing and
saving as draft at the end of a flow, so I've renamed it to
'complete_workflow'.