Add a service to extract the redirect location from form requests
This accepts a URL, a request body and a method. It then makes
a request to the url with the request body and then, if the response
is a redirect, it returns the location. This will allow us to intercept
requests to smart answer forms and render the page within the preview.

There’s some safeguards here to ensure the response is one we expect,
and that the URL is a GOV.UK one.