Configure Prometheus Exporter
The motivation for this change was that I was getting a lot of errors
logged locally due to a recent change in `govuk_app_config` as it
couldn't find a locally running Prometheus instance.

This change just adds some initializer code to make sure the app is
configured properly for Prometheus. See examples here:

https://github.com/alphagov/frontend/blob/main/config/initializers/prometheus.rb
https://github.com/alphagov/whitehall/blob/main/lib/collectors/scheduled_publishing_collector.rb

This will also mean that we can define our own custom metrics to track
in Prometheus if needed in future.