Update edition types for older versions
Before replatforming, all our models were namespaced
under the `ContentBlockManager::ContentBlock` namespace.

When we replatformed, we moved the data over manually. However,
we forgot to update the `item_type` values for older versions. This
resulted in versions that were created before the replatforming to
not show up.

This adds a simple Data Migration to update all the `item_type`s with
the old value to the new, non-namespaced value. I’ve tested this in
development with data from production and it works perfectly!