Handle already published state
It is possible for a user to have multiple tabs with identical changes, if they publish both tabs this currently throws an error, we do not need it to so this updates the behaviour to allow the flow to continue if the entry has already been published in a different tab.

We only need the PublishEditionService to be called for unpublished entries, but we want both published and unpublished state entries to path to the confirmation page.