Fix flaky test
Ocassionally there are a couple of tests that fail when tests are run in
a particular order on `spec/requests/documents_spec.rb`.

After a bit of detective work, I’ve managed to work out that it’s this
test that causes the issue, because the test strategy gets changed for
all subsequent tests.

To get round this, we just mock the response from
`Flipflop.show_all_content_block_types?`. As mocks get reset after each
test, there is no remaining state hanging round and polluting the
test environment.