Remove attribute references from embed codes when searching
When searching by an embed code for a specific field, we get no results
because the only referenced embed code is the root embed code. Rather
than storing every single possible embed code in the database, we check
if the keyword filter is an embed code, and if it is, we remove any
attribute references and ONLY search for the root embed code.

I’ve also added a Cucumber spec to test this functionality. I’ve had to
change the setup step too, so `sluggable_string` uses a `-`, rather than
an underscore, as underscores are not used in content blocks in the
application, so the embed codes created in test are not valid.