Adjust description of editing draft (creates new)
If you choose to edit a draft edition, for example by:

- start creating an edition
- return to homepage
- go to the draft edition
- select "Edit edition"

this current *creates* a further new edition, rather than
edit the existing draft. Is this correct behaviour or
something we should change in the future? It might be rather
frustrating for the user if they've invested a lot of effort
in a complex draft and are unable to finish off the job of
creating that new edition...