Convert DefaultBlockComponent test to RSpec
Note that this test describing the
`Document::Show::DefaultBlockComponent` was incorrectly placed
in the `components/edition/show` directory.

In the future we will probably move this into the "edition"
scope/directory as it's now clearly concerned with
representing the default block for a given edition, rather
than a particular document.