Add PublishedEdition controller
This is very similar to the documents controller, except it
renders the latest published edition rather than the most
recent edition.

Let hold off on "drying" this up to a single controller
until we understand better whether/how the view of a
published edition may need to differ from that of the most
recent edition.