Fix: show published content for embedded objects
Seen when using `BlocksComponent`, e.g. to show "Telephones"
tab to view the published edition (`/{DocumentId}/published_edition`).

This component is invoked by the `SubschemaItemComponent`
for each "embedded object".

The component was taking a `Document` and then displaying
content for the most recent edition. In this commit we
change the `BlocksComponent` to take an `Edition` instead.