Remove remaining uses of Edition foreign keys
We remove all references to these 2 foreign keys on the
`documents` table, which are no longer required:

1. `documents.latest_edition_id`: we now get the
`Document#latest_published_edition` and
`Document#most_recent_edition` through the Edition model,
using the `Edition#state` to understand whether it's published.

2. `documents.live_edition_id`: the `Document.live` scope
now purely joins on `editions` to find the most recent
editions in the published state

In a future PR, once this is released, we'll remove the 2
columns from the `documents` table.