Conditionally render location link based on edition state.
If the edition is in any state other than published we want to
be able to preview the embedded block, this should apply for 'draft'
and 'awaiting 2i' currently, but will also need to account for 'factcheck'
in the future.

If the edition is in published state we instead want the title to link to
the page it is embedded in, and to not show the preview link at all, since
the version can be seen live.