Convert app/workers tests to rspec
Once I had converted the publish_intent_worker file, it began to fail
for an unrelated reason. It seems the PublishIntentPresenter was
updated to only take two arguments (probably before or during the
migration to this repo). As such, RSpec correctly failed the tests as
we were still passing three arguments. I updated the one place in the
code where we're using PublishingIntentPresenter to only pass two
args and once in a spec file also.