Ensure deleted editions aren't scheduled
I intended to implement the behaviour for this here however I realised
that the SchedulePublishingWorker already ensures deleted editions
aren't scheduled. The only status that it allows is :scheduled so any
edition in the :deleted state would be ignored.

As such this commit just describes the behaviour I wanted as a test.