Show review outcome in timeline
If the version has the `awaiting_factcheck` state (which it
gets when its created on the state transition, see
`HasAuditTrail`) then we show some information about the
"review" process. We've captured this before transitioning
to "awaiting_factcheck".

The "review outcome" consists of 3 pieces of information:

1. `editions.review_skipped` - was the review performed or
    skipped?

2. `editions.review_outcome_recorded_by` - who is giving us
   this information about the review?

3. `editions.review_outcome_recorded_at` - when was this
    information provided?

At present, in the timeline entry under "Sent to factcheck"
we only show whether the review was skipped or performed.
This is because both:

- the timestamp of when the outcome was recorded
- and the identity of the person recording the review
outcome

are the same as the timestamp and "actor" of the "Sent to
factcheck" event within which the review outcome information
is contained.