Record completion on Workflow::UpdateMethods#complete_workflow
When reviewing a draft the editor is required to confirm
that the contents of the draft block are accurate. Once
this is done we proceed to "complete the workflow" by
running `WorkflowCompletion#call`. We record the completeion
as a timestamp on `editions.workflow_completed_at` and
will test its presence with `Edition#completed?`.