Extend WorkflowCompletion to handle 'send_to_review'
On the workflow "review" step (`/editions/{editionId}/workflow/review`)
we need to:

1. ensure that the editor has checked the content
2. record the fact that the workflow is complete
  (`editions.workflow_completed_at`)

before attempting the `Edition#ready_for_review` transition.

Currently the "review" form on that page has a submission
button POSTing to the `StatusTransitionsController`. However,
as we need to verify that the workflow is complete it makes
more sense for this form to always POST to
the `WorkflowController`. The controller can then pass the
"action" of "send_to_review" to the WorkflowCompletion
class which is easily extended to handle another type of
"completion".

How the "Send to review" call to action needs behave
differently depending on location:

1. WORKFLOW REVIEW PAGE
- to change: "Send to review"
  -> POST to `WorkflowController`

2. SHOW PAGE
- to leave unchanged: Draft complete: "Send to review"
  -> POST to `StatusTransitionsController`

- to change: Draft incomplete: "Resume draft"
  -> GET to workflow review (`/editions/{editionId}/workflow/review`)