Remove proposed publish date from scheduled status
Prior to this, the date in the status row referred to the date the
user moved the edition into the status, with the exception of the
Scheduled status where this showed the proposed publish date.

This commit swaps removes the scheduled date from this row as it now
has its own row underneath, and just shows the user name.

It could also show the date that the user moved this into the
scheduled state but it could be confusing for the user and all the
information is presented in the timeline if they're curious.