Offer editor of incomplete draft link to 'Complete draft'
Rather than the "Send to 2i" call-to-action which is only
appropriate if the draft workflow has been completed.

It's on the workflow review step (`/editions/ {editionId}/workflow/review`)
that the editor has the opportunity to review all the
content in the draft and confirm that it's been checked.
On confirmation we set the `Edition.workflow_completed_at`
timestamp.