Require draft to be completed before transition to 'awaiting_review'
We add a "guard" to the `:awaiting_review` transition to
check that the draft workflow has been completed.

We also update the specs for the StatusTransitionsController
-- this controller can still perform transitions to
"awaiting review" when receiving a POST from a `Send to 2i`
call to action on the edition show page. This is offered only
if the draft edition is `completed?`.