Move 'mark complete' to WorkflowCompletion
We have now settled on using "workflow completion"
rather than "content checked" as our guard on the transition
from "draft" to "awaiting review". It therefore makes sense
for the `WorkflowCompletion` class to be responsible for
recording the fact that the edition has completed the
drafting workflow (by setting the `editions.workflow_completed_at`
timestamp).

We refactor the `WorkflowCompletion` a little so that we
raise `UnhandledSaveActionError` early if the requested
"save action" is unknown -- and only mark the edition as
completed if the save action is valid.