Add failing test to illustrate the bug
The issue we have is that when we click “Edit” on a draft document that
has no published editions, we’re creating a new document.

This is because if an edition has no published documents, we assume
we’re creating a new edition. This results in a brand new document
being  created for the edition, which essentially copies the basic
information (Title etc) from the old edition, and creates a new
edition AND document.