Build 'Delete draft' button with DELETE method
It was being built using the PATCH method, resulting in a
"Can't verify CSRF token authenticity." error.

This code was brought over from Whitehall so may never have
worked.

It's frustrating that the error can't be reproduced in a
cucumber feature, so instead we verify that the form has the
correct HTML attributes.