Allow nested field configuration to work correctly
When creating nested fields, we pass the config of the parent field to
the schema to ensure the configuration is read correctly. I’ve also made
the `Schema::Field#config` method public, so we can unit test it more
easily.

This also adds govspeak config for the contact link description field,
which was working by accident before.