'Cancel' creation of new embedded object using back_link
Previously the "Cancel" link on the "new embedded object" form
(`editions/embedded_objects/new.html.erb`) was linking to
`/editions/{editionId}/workflow/embedded_{subschema.id}` e.g.
`/editions/{editionId}/workflow/embedded_contact_links`.

This ought to be a valid route, but in fact was raising a

`Step embedded_contact_links does not exist (ActionController::RoutingError)`

from [`Workflow::HasSteps`][]

This is because, in this scenario, when the `Workflow::Steps`
class prepares a list of available steps as candidates for
`WorkflowController#show` to use in [`current_step`][],
it doesn't include our subschemas.

This workflow step code is doing rather too much and needs
refactoring to make it simpler and easier to change. Whilst
we think about how best to do that, in this commit we introduce
a simple fix/work-around for the current bug.

Rather than redirect to the workflow step for the particular
type of embedded object we now redirect to the link supplied
to the "new" form's "back link", which is:

`/editions/{editionId}/workflow/group_contact_methods`

This returns the editor to the view of:

- all the block's contact methods (if any exist)
- the list of radio buttons to create a first contact method
  (if none exist yet)

[`Workflow::HasSteps`]:
https://github.com/alphagov/content-block-manager/blob/79097f8141ea5844908a1639646b99c5d92e90b5/app/controllers/concerns/workflow/has_steps.rb#L34-L36

[`current_step`]:
https://github.com/alphagov/content-block-manager/blob/a8a9f20add3efe1480483fca1f0239b56013465e/app/controllers/editions/workflow_controller.rb#L8-L12