Allow objects to be conditionally hidden
If a field has a `show_field`, then we render this as a checkbox
with the remaining fields hidden within the conditional wrapper.

This also means we can get rid of the “special” components to
render things like Call Charges etc.

There are a couple of TODOs here, because the translations are broken
for some fields. Rather than making this PR even bigger, I’ve added a
TODO and will fix it in the next PR.