Provide guidance on URL format in hint text
In addition to omitting the "https://" prefix there are
several ways in which the format of the url could fail
validation (see the [schema's regex][]).

Because we don't record exactly how the given string failed
to validate, it's more appropriate to put our guidance in
a "hint" element rather than include it as part of the
validation error message.

[schema's regex]:
https://github.com/alphagov/publishing-api/blob/main/content_schemas/formats/content_block_contact.jsonnet#L161-L163