Create new Outcome model
This will eventually replace the extra fields we added to Edition
which capture both the Review (2i) Outcome and the Factcheck Outcome.

Creates a base Outcome model and two subclassed models utilising
single table inheritance. This means we can share the db fields for
each and don't need to create two tables.

The slight quirk here is that the factcheck review is currently
storing the reviewer as a name/email string and the 2i reviewer is
storing the user ID of the 2ier but we can work with this and it's
resolved in future commits.